+++ examples/src/fireworks/src/fireworks.pax
@@ -1,6 +1,6 @@
-<Group id=hello @wheel=self.handle_scroll>
+<Group @wheel=self.handle_scroll >
     for i in 0..60 {
-        <Rectangle class=rect  width=300px height=300px />
+        <Rectangle class=rect width=300px height=300px />
     }
 </Group>
 
@@ -10,23 +10,25 @@
 
 @settings {
     .rect {
-        fill: {hsl((ticks + i * 360.0 / 30.0)%, 75%, 75%)}
-        anchor_x: 50%,
-        anchor_y: 50%,
-        rotate: {(i + 2) * rotation + ticks / 1000.0},
-        scale_x: {
-            (
-                100*
-                (0.75 + (i * rotation)) *
-                (1 - ((rotation / 5) + i / 1000.0))
-            )%
-        },
-        scale_y: {
-            (
-                100*
-                (0.75 + (i * rotation)) *
-                (1 - ((rotation / 5) + i / 1000.0))
-            )%
-        }
+        //fill: {Fill::Solid(hsl((ticks + i * 360.0 / 30.0)%, 75%, 75%))},
+        //rotate: {((i + 2) * rotation + ticks / 1000.0)deg},
+        //scale_x: {
+        //    (
+        //        100*
+        //        (0.75 + (i * rotation)) *
+        //        (1 - ((rotation / 5) + i / 1000.0))
+        //    )%
+        //},
+        //scale_y: {
+        //    (
+        //        100*
+        //        (0.75 + (i * rotation)) *
+        //        (1 - ((rotation / 5) + i / 1000.0))
+        //    )%
+        //},
+        //anchor_x: 50%,
+        //anchor_y: 50%,
+        x: {50%},
+        y: 50%,
     }
 }
